+++ zcms/src/com/mavict/article/image/ArticleImageService.java
@@ -11,7 +11,7 @@ import com.mavict.base.BaseService;
  * @author 沧海软件(北京)有限公司
  * @date 2016年1月16日 上午10:59:19
  */
-public interface ArticleImageService extends BaseService<ArticleImage, Long> {
+public interface ArticleImageService extends BaseService<ArticleImage, Integer> {
 	/**
 	 * 获取Article上传文章图片
 	 *
