+++ zcms/src/com/mavict/guestbook/GuestbookDao.java
@@ -10,7 +10,7 @@ import com.mavict.base.BaseDao;
  * @author 沧海软件(北京)有限公司
  * @date 2015年10月31日 下午8:36:56
  */
-public interface GuestbookDao extends BaseDao<Guestbook, Long> {
+public interface GuestbookDao extends BaseDao<Guestbook, Integer> {
 	/** 回复留言 */
-	void reply(Long id,HttpServletRequest request);
+	void reply(Integer id,HttpServletRequest request);
 }
